Moved a few classes into eva-accession-core as suggested by Cristina
