+++ eva-accession-core/src/test/resources/input-files/assembly-report/AssemblyReport.txt
@@ -1,0 +1,67 @@
+# Assembly name:  GRCm38.p6
+# Description:    Genome Reference Consortium Mouse Build 38 patch release 6 (GRCm38.p6)
+# Organism name:  Mus musculus (house mouse)
+# Infraspecific name:  strain=C57BL/6J
+# Taxid:          10090
+# BioProject:     PRJNA20689
+# Submitter:      Genome Reference Consortium
+# Date:           2017-9-15
+# Assembly type:  haploid-with-alt-loci
+# Release type:   patch
+# Assembly level: Chromosome
+# Genome representation: full
+# RefSeq category: Reference Genome
+# GenBank contigSynonyms accession: GCA_000001635.8
+# RefSeq contigSynonyms accession: GCF_000001635.26
+# RefSeq contigSynonyms and GenBank assemblies identical: no
+#
+## Assembly-Units:
+## GenBank Unit Accession	RefSeq Unit Accession	Assembly-Unit name
+## GCA_000000055.2	GCF_000000055.19	Primary Assembly (C57BL/6J)
+## GCA_000291555.6	GCF_000291555.6	PATCHES
+## GCA_000000495.1	GCF_000000495.3	129P2/OlaHsd
+## GCA_000001665.1	GCF_000001665.3	129S1/SvImJ
+## GCA_000001675.1	GCF_000001675.3	129S2/SvPas
+## GCA_000000395.2	GCF_000000395.5	129S6/SvEvTac
+## GCA_000237955.1	GCF_000237955.1	129X1/SvJ
+## GCA_000000325.1	GCF_000000325.5	A/J
+## GCA_000000425.1	GCF_000000425.1	AKR/J
+## GCA_000000335.2	GCF_000000335.9	BALB/c
+## GCA_000000465.2	GCF_000000465.4	CAST/Ei
+## GCA_000001715.1	GCF_000001715.3	NOD/MrkTac
+## GCA_000001705.1	GCF_000001705.1	NOD/ShiLtJ
+## GCA_000001725.1	GCF_000001725.3	RIII
+## GCA_000000385.2	GCF_000000385.11	UNKNOWN
+## GCA_000001685.1	GCF_000001685.3	129S7/SvEvBrd-Hprt-b-m2
+## GCA_000004175.1	GCF_000004175.1	non-nuclear
+#
+# Ordered by chromosome/plasmid; the chromosomes/plasmids are followed by
+# unlocalized scaffolds.
+# Unplaced scaffolds are listed at the end.
+# RefSeq is equal or derived from GenBank object.
+#
+# Sequence-Name	Sequence-Role	Assigned-Molecule	Assigned-Molecule-Location/Type	GenBank-Accn	Relationship	RefSeq-Accn	Assembly-Unit	Sequence-Length	UCSC-style-name
+chrom1	assembled-molecule	1	Chromosome	CM000994.2	=	NC_000067.6	C57BL/6J	195471971	chr1
+2	assembled-molecule	2	Chromosome	CM000995.2	=	NC_000068.7	C57BL/6J	182113224	chr2
+3	assembled-molecule	3	Chromosome	CM000996.2	=	NC_000069.6	C57BL/6J	160039680	chr3
+4	assembled-molecule	4	Chromosome	CM000997.2	=	NC_000070.6	C57BL/6J	156508116	chr4
+5	assembled-molecule	5	Chromosome	CM000998.2	=	NC_000071.6	C57BL/6J	151834684	chr5
+6	assembled-molecule	6	Chromosome	CM000999.2	=	NC_000072.6	C57BL/6J	149736546	chr6
+7	assembled-molecule	7	Chromosome	CM001000.2	=	NC_000073.6	C57BL/6J	145441459	chr7
+X	assembled-molecule	X	Chromosome	CM001013.2	=	NC_000086.7	C57BL/6J	171031299	chrX
+Y	assembled-molecule	Y	Chromosome	CM001014.2	=	NC_000087.7	C57BL/6J	91744698	chrY
+MMCHR1_RANDOM_CTG1	unlocalized-scaffold	1	Chromosome	GL456210.1	=	NT_166280.1	C57BL/6J	169725	chr1_GL456210_random
+CHR_without_synonym	unlocalized-scaffold	1	Chromosome	GL_without_synonym	<>	NT_without_synonym	C57BL/6J	241735	UCSC_without_synonym
+MMCHR1_RANDOM_CTG5	unlocalized-scaffold	1	Chromosome	GL456221.1	=	NT_162750.1	C57BL/6J	206961	chr1_GL456221_random
+MMCHR4UN_CTG1	unlocalized-scaffold	4	Chromosome	GL456216.1	=	NT_166291.1	C57BL/6J	66673	chr4_GL456216_random
+MMCHR4UN_CTG3	unlocalized-scaffold	4	Chromosome	GL456350.1	=	NT_166434.1	C57BL/6J	227966	chr4_GL456350_random
+MMCHR4UN_CTG2	unlocalized-scaffold	4	Chromosome	JH584292.1	=	NT_187052.1	C57BL/6J	14945	chr4_JH584292_random
+MMCHR5_RANDOM_CTG4	unlocalized-scaffold	5	Chromosome	GL456354.1	=	NT_166438.1	C57BL/6J	195993	chr5_GL456354_random
+MMCHR5_RANDOM_CTG3	unlocalized-scaffold	5	Chromosome	JH584298.1	=	NT_187058.1	C57BL/6J	184189	na
+MMCHR5_RANDOM_CTG5	unlocalized-scaffold	na	Chromosome	JH584299.1	=	NT_187059.1	C57BL/6J	953012	chr5_JH584299_random
+MMCHR7_RANDOM_CTG1	unlocalized-scaffold	7	Chromosome	GL456219.1	=	NT_166307.1	C57BL/6J	175968	chr7_GL456219_random
+MMCHRX_RANDOM_CTG2	unlocalized-scaffold	X	Chromosome	GL456233.1	=	NT_165789.2	C57BL/6J	336933	chrX_GL456233_random
+22	assembled-molecule	22	Chromosome	genbank_example_1	=	refseq_example_1	C57BL/6J	91744698	ucsc_example_1
+otherprefix_chr45	assembled-molecule	45	Chromosome	genbank_example_2	=	refseq_example_2	C57BL/6J	91744698	ucsc_example_2
+sequence_name_example_3	assembled-molecule	assigned_molecule_example_3	Chromosome	genbank_example_3	=	refseq_example_3	C57BL/6J	91744698	ucsc_example_3
+sequence_name_example_4	assembled-molecule	assigned_molecule_example_4	Chromosome	genbank_example_4	=	NW_006738765.1	C57BL/6J	91744698	ucsc_example_4
