+++ eva-accession-import/pom.xml
@@ -20,7 +20,6 @@
         <dependency>
             <groupId>uk.ac.ebi.eva</groupId>
             <artifactId>eva-accession-core</artifactId>
-            <version>0.3.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>uk.ac.ebi.eva</groupId>
