+++ eva-accession-import/src/test/java/uk/ac/ebi/eva/accession/dbsnp/processors/SubmittedVariantRenormalizationProcessorTest.java
@@ -20,9 +20,9 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import uk.ac.ebi.eva.accession.core.persistence.DbsnpSubmittedVariantEntity;
-import uk.ac.ebi.eva.accession.dbsnp.contig.ContigMapping;
-import uk.ac.ebi.eva.accession.dbsnp.contig.ContigSynonyms;
-import uk.ac.ebi.eva.accession.dbsnp.io.FastaSynonymSequenceReader;
+import uk.ac.ebi.eva.accession.core.contig.ContigMapping;
+import uk.ac.ebi.eva.accession.core.contig.ContigSynonyms;
+import uk.ac.ebi.eva.accession.core.io.FastaSynonymSequenceReader;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
