+++ eva-accession-pipeline/src/test/resources/accession-pipeline-test.properties
@@ -13,8 +13,8 @@ parameters.vcf=src/test/resources/input-files/vcf/small_genotyped.vcf.gz
 parameters.vcfAggregation=NONE
 
 parameters.fasta=src/test/resources/input-files/fasta/Homo_sapiens.GRCh37.75.chr20.head_1200.fa
-parameters.outputVcf=/tmp/accession-output.vcf
-#(Windows path) parameters.outputVcf=C:/Users/(user)/AppData/Local/Temp/accession-output.vcf
+#parameters.outputVcf=/tmp/accession-output.vcf
+parameters.outputVcf=C:/Users/sundarvenkata/AppData/Local/Temp/accession-output.vcf
 
 spring.jpa.show-sql=true
