+++ eva-accession-release/pom.xml
@@ -28,7 +28,7 @@
         </dependency>
         <dependency>
             <groupId>uk.ac.ebi.eva</groupId>
-            <artifactId>eva-accession-import</artifactId>
+            <artifactId>eva-accession-core</artifactId>
         </dependency>
         <dependency>
             <groupId>com.github.samtools</groupId>
