+++ pom.xml
@@ -32,7 +32,7 @@
         <dependencies>
             <dependency>
                 <groupId>uk.ac.ebi.eva</groupId>
-                <artifactId>eva-accession-import</artifactId>
+                <artifactId>eva-accession-core</artifactId>
                 <version>0.3.0-SNAPSHOT</version>
             </dependency>
             <dependency>
@@ -48,12 +48,12 @@
             <dependency>
                 <groupId>uk.ac.ebi.eva</groupId>
                 <artifactId>variation-commons-core</artifactId>
-                <version>0.6</version>
+                <version>0.7-SNAPSHOT</version>
             </dependency>
             <dependency>
                 <groupId>uk.ac.ebi.eva</groupId>
                 <artifactId>variation-commons-batch</artifactId>
-                <version>0.6</version>
+                <version>0.7-SNAPSHOT</version>
             </dependency>
 
             <dependency>
